prepare setup for mongodb physical backups
this updates the setup to cater for mongodb physical backup requirements such as:
- mongod nodes being shutdown during restores
- pbm-agents need read/write access to mongodb data path.
To cover for those, we create 3 separate containers for the pbm-agents.
That way, they stay alive after the mongod containers are shutdown.

We also create an extra container for pmm-client so that it stays alive
after both mongod and pbm-agent containers are shutdown.